% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseme.R
\name{praise_me}
\alias{praise_me}
\title{Praises user}
\usage{
praise_me(name)
}
\arguments{
\item{name}{User-specified string, typically a name of a person.}
}
\value{
A string to praise the name
}
\description{
\code{praise_me} returns praise incorporating a user-specified string.
}
\examples{
praise_me("Bob")
}
