% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{praise_messages}
\alias{praise_messages}
\title{Example Praise Messages}
\format{
\subsection{\code{praise_messages}}{

A character vector of length 3:
\itemize{
\item "Amazing!"
\item "Fantastic!"
\item "Brilliant!"
}
}
}
\source{
Created for teaching
}
\usage{
praise_messages
}
\description{
A toy dataset of encouraging messages.
}
\keyword{datasets}
