% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praiseJohn}
\alias{praiseJohn}
\title{Delivering praise to John}
\usage{
praiseJohn()
}
\value{
Text string with praise to John
}
\description{
This function is useful to praise John when he is sad.
}
\examples{
praiseJohn()
}
