% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseme.R
\name{praiseme}
\alias{praiseme}
\title{Praises user}
\usage{
praiseme(name)
}
\arguments{
\item{name}{user-specified name}
}
\value{
a character string
}
\description{
\code{praiseme} returns praise for the name given
}
\examples{
praiseme("Alex")
}
