% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseme_graphics.r
\name{praiseme_graphics}
\alias{praiseme_graphics}
\title{Display a short message of praise to the user}
\usage{
praiseme_graphics(praisefor, colour1 = "black", colour2 = "white")
}
\arguments{
\item{praisefor}{The verb or subject to praise the user for, entered as a string. If blank, a generic praise message will be returned.}

\item{colour1}{The color of the text. Default is black.}

\item{colour2}{The color of the background. Default is white.}
}
\value{
A short praise message displayed as a plot.
}
\description{
Takes a string as input and displays a short message of praise using the string.
}
\examples{
"art"
"cooking"
"writing"

}
