% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praiseWithPunc}
\alias{praiseWithPunc}
\title{Delivering praise to Name with punctuation}
\usage{
praiseWithPunc(x = "John", punctuation = "!")
}
\arguments{
\item{x}{This is the name of the person the function will praise.}

\item{punctuation}{empasis type}
}
\value{
Text string with praise to Name
}
\description{
This function is useful to praise Name when they are sad.
}
\examples{
praise()
}
