% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praiseYou}
\alias{praiseYou}
\title{Praising a person}
\usage{
praiseYou(name, punctuation)
}
\arguments{
\item{name}{This is the name of the person you want to praise}

\item{punctuation}{This is the punctuation at the end of the praise
message.}
}
\value{
returns a nice message with a praise name
}
\description{
Now this function can be personilised!
}
\examples{
praiseYou(name = "Paula", punctuation = "!") # will give you a nice message with the given name
}
