% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{Your name}

\item{punctuation}{Your punctuation}
}
\value{

}
\description{
Praise
}
\examples{
praise("Aidan", "!")
}
