% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; containing a name}

\item{punctuation}{text string; the symbol to end the message with following the name.}
}
\value{
text string with a phrase containing the name and punctuation
}
\description{
This function returns a string of praise.
}
\examples{
praise(name = "Steph", punctuation = "!")
}
