% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Send Praise}
\usage{
praise(name = "Alicia")
}
\arguments{
\item{name}{character, the name of the person you want to praise}
}
\description{
This function provides the user with necessary and appreciated praise.
Yay praise!
}
\examples{
praise()
}
