% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation)
}
\arguments{
\item{name}{This is the name of the person I want to praise. Input as text string.}

\item{punctuation}{This is our emphasis as a text input.}
}
\value{
Text string with praise
}
\description{
This function is super useful when you're feeling sad and need praise.
}
\examples{
praise(name = "Alyce", punctuation = "!")
}
