% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; who do you want to be praised?}

\item{punctuation}{text string; emphasis}
}
\value{
text string of praise
}
\description{
Provides some praise here and there
}
