% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person I want to praise}

\item{punctuation}{text string; This is our emphasis as a text input}
}
\value{
text string with praise
}
\description{
Deliver Praise
}
\details{
This function is super useful when you're feeling sad
}
\examples{
praise(name = "Steph", punctuation = "!")
}
