% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Delivers praise to the user}
\usage{
praise(name = "Christy")
}
\arguments{
\item{name}{Name of the person you wnat to prasie.
Must be a character. Default's to the name, "Christy".}
}
\value{
a nice message to the user
}
\description{
Delivers praise to the user
}
\examples{
praise(name = "Christy")
}
