% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string: this is the name of the person for praising}

\item{punctuation}{text string: this is the emphasis as a text input}
}
\value{
text string with praise
}
\description{
This function is made to cheer you up
}
\examples{
praise(name = "Steph", punctuation = "!")
}
