% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name = "AAAAAAAAA", punctuation = "!!!!!!!")
}
\arguments{
\item{name}{Name of the praise recipient. Requires character input}

\item{punctuation}{Desired punctuation}
}
\value{
Text string with praise
}
\description{
This function is super useful for returning praise.
}
\examples{
praise(name = "AAAAAAAAA",
       punctuation = "!!!!!!!")
praise("AAAAAAAAA")

}
