% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{praise help}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person who I want to praise}

\item{punctuation}{text string; The desired punctuation as a text input}
}
\value{
Text string
}
\description{
This function reminds that you're the realest and nobody can f*ck with you.
}
\examples{
praise(name = "POP SMOKE", punctuation = "?")
}
