% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{praise}
\usage{
praise(name = "default", punctuation = "!")
}
\arguments{
\item{name}{Name of person you want to praise}

\item{punctuation}{the punctuation after the name}
}
\value{
text string
}
\description{
praise
}
\examples{

}
