% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise someone}
\usage{
praise(name = "Gaby", adjective = "Best", punctuation = "!")
}
\arguments{
\item{name}{name of praise recipient}

\item{adjective}{type of praise you want to send}

\item{punctuation}{sets the tone of the praise}
}
\description{
giving people praise
}
\examples{
praise("Broccoli")
}
