% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Delivery Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person you want to praise}

\item{punctuation}{text string; this is our emphasis as a text input.}
}
\value{
Text string with praise
}
\description{
This function is great to cheer you up when you are feeling down.
}
\examples{
praise(name = "Jadyn", punctuation = "!" )
}
