% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Recieve praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person I want to praise}

\item{punctuation}{text string; This is our emphasis as a text input.}
}
\value{
Text string with praise
}
\description{
Recieve praise
}
\examples{
praise(name = "Aaron", punctuation = "!")
}
