% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Send Some Praise}
\usage{
praise(name = "Jessie")
}
\arguments{
\item{name}{character, the name of a person you want to praise}
}
\description{
Sends love to someone in the form of praise.
}
\examples{

# Praise the lord.
praise()

}
