% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Delivering praise to Name}
\usage{
praise(x = "John")
}
\arguments{
\item{x}{This is the name of the person the function will praise.}
}
\value{
Text string with praise to Name
}
\description{
This function is useful to praise Name when they are sad.
}
\examples{
praise()
}
