% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Give a user some praise}
\usage{
praise(user_name)
}
\arguments{
\item{user_name}{The name of a user as a character vector}
}
\value{
The \code{user_name} with the words "You're amazing" preprended to it.
}
\description{
This function gives the user a positive affirmation, because they deserve it!
}
\examples{
praise("Stuart")
praise("Justin")

}
