% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseme.R
\name{praise}
\alias{praise}
\title{Send Some Praise}
\usage{
praise(name = "Justin")
}
\arguments{
\item{name}{character, the name of the person that deserves somse praise}
}
\description{
Praise someone at critical times in their life
}
\examples{

# yay you're about to get some praise!
praise()
}
