% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Delivers praise}
\usage{
praise(name, punctuation)
}
\arguments{
\item{name}{Text string of the name I would like to praise}

\item{punctuation}{This is our punctionation as a text input}
}
\value{
Returns text string with praise
}
\description{
This function is useful when you are
feeling sad and your name is Kristy.
}
\examples{
praise(name = "Kristy", punctuation="!")

}
