% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person I want to praise}

\item{punct}{text string; This is our emphasis as a text input}
}
\value{
Text string with praise
@export

@examples
praise(name = "Steph", punctuation = "!")
}
\description{
This function is super useful when you're feeling sad and is useful for everyone.
}
