% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation)
}
\arguments{
\item{name}{This is the name of person I want to praise. Input as text strong}

\item{punctuation}{This is our emphasis as a text input.}
}
\value{
Text string with praise
}
\description{
This function is super useful when you're feeling sad
}
\examples{
praise(name = "Steph", punctuation = "!")
}
