% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise someone}
\usage{
praise(name = "Laura", punctuation = "!")
}
\arguments{
\item{name}{specifies recipient of praise}

\item{punctuation}{Specifies tone of praise}
}
\description{
Let's dole out praise
}
\examples{
praise("Nick")
praise("Holly")

}
