% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; Name of person I wish to praise.}

\item{punctuation}{text string; The punctuation provided at the end of the statement.}
}
\value{
Text string with praise
}
\description{
Deliver Praise
}
\examples{
praise(name = "Steph", punctuation = "!")
}
