% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{This is the name of the person I want to praise. Input as text string}

\item{punctuation}{This is our emphasis. Input as text}
}
\value{
Text string with praise
}
\description{
This function is useful when you feel sad
}
\examples{
praise(name = "Steph", punctuation = "!")
}
