% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.r
\name{praise}
\alias{praise}
\title{Praise}
\usage{
praise(name, punctuation)
}
\arguments{
\item{name}{Who to praise}

\item{punctuation}{What to end in.}

\item{potato}{why not}
}
\value{
Text string with praise
}
\description{
make some praise options.
}
\examples{
praise()

}
\author{
Dale Maschette
}
