% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priase.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person to praise}

\item{punctuation}{text string; This is the emphasis as a text input}
}
\value{
Text string with praise
}
\description{
A very nice little function that gives you some praise when you're feeling down.
Please be mindful that overuse can lead to undesirable effects.
}
\examples{
praise(name = "Mathew", punctuation = "!")
}
