% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person in need of praise.}

\item{punctuation}{text string; This is our emphasis as a text input.}
}
\value{
Text string with praise.
}
\description{
When you've got the blues and need to be reminded who's boss.
}
\examples{
praise(name = "Steph", punctuation = "!")
}
