% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{This is the name of the person you want to praise}

\item{punctuation}{This is our emphasis as a text input}
}
\value{
with praise
}
\description{
This function is super useful during sad times.
}
