% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name)
}
\arguments{
\item{name}{This is the name of the person I want to praise.}
}
\value{
Text string with praise
}
\description{
This function is super useful when you're feeling sad and Nick
needs praise.
}
\examples{
praise("Steph")
}
