% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Show Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; The name of the person whom you wish to receive praise}

\item{punctuation}{text string; this is the exmphasis}
}
\value{
Text string with praise
}
\description{
For when you believe someone is feeling sad and is deserving of praise
}
