% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name = "Mitch", punctuation = "!")
}
\arguments{
\item{name}{This is the name of the person I want to praise. Defaults to Mitch. Input as text string.}

\item{punctuation}{This is our emphasis as a text string. Defaults to !}
}
\value{
Text string with praise
}
\description{
This function is super useful when your feeling sad.
}
\examples{
praise()
}
