% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise someone}
\usage{
praise(name = "Saskia")
}
\arguments{
\item{name}{Specifies recipient of praise}
}
\description{
Let's dole out praise
}
\examples{
praise("Robbie")
praise("Saskia")
}
