% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{A character string; the name of the person to praise}

\item{punctuation}{A character string; emphasis for the praise}
}
\value{
A character string with praise
}
\description{
This function is for when you're feeling sad
and is useful for everyone
}
\examples{
praise("Thomas")
praise("Alice","!!!!")

}
