% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{praise message}
\usage{
praise()
}
\value{
returns a nice message with praise
}
\description{
Praises you with a sweet message.
}
\examples{
praise() # will give you a nice message
}
\keyword{praise}
