% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; Name of person to be praised}

\item{punctuation}{text string; This is our emphasis}
}
\value{
Text string with praise
}
\description{
This function is super useful when wanting praise
}
\examples{
praise(name = "Steph", punctuation = "!")
}
