% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.r
\name{praise}
\alias{praise}
\title{Praise someone}
\usage{
praise(name = "Rhydwyn")
}
\arguments{
\item{name}{Specifies the recipient of praise}
}
\description{
dole out the praise ❤️
}
\examples{
praise("John")
}
