% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise someone}
\usage{
praise(name = "Robbie", verb = "loves")
}
\arguments{
\item{name}{Specifies recipient of name}

\item{verb}{Specififies how saskia feels}
}
\description{
Let's dole out praise
}
\examples{
praise("Anna")
praise("Robbie","hates")
}
