% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.r
\name{praise}
\alias{praise}
\title{Praise someone}
\usage{
praise(name = "sadia", exclaim = "!")
}
\arguments{
\item{name}{Specifies recipient of praise}

\item{exclaim}{Specifies the tone of voice}
}
\description{
Let's dole out praise
}
\examples{
praise('Nick')
praise('Holly')

}
