% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise gives praise}
\usage{
praise(name = "Steve")
}
\arguments{
\item{name}{The name of the person to be praised.}
}
\description{
This function returns a message of praise when invoked.
}
