% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{praise}
\alias{praise}
\title{Send some praise}
\usage{
praise(name = "Thilaksha")
}
\arguments{
\item{name}{character, the name of the person you wan to praise}
}
\description{
Praise someone at critical times in my life
}
\examples{

# I am gonna deliver some praise!

}
