% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Title}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; the name of the person to be praised.}

\item{punctuation}{text string; the emphasis on text.}
}
\value{
text string with praise
}
\description{
This functions is useful when you're feeling a little down.
}
\examples{
praise(name = "Heather", punctuation = "!")

}
