% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation)
}
\arguments{
\item{name}{Name of person to praise.}

\item{punctuation}{This is our emphasis as a text output.}
}
\value{
Text string with praise
}
\description{
Useful when you are down.
}
\examples{
praise(name="Steph",punctuation="!")
}
