% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name)
}
\arguments{
\item{name}{text string, Name of person you want to praise}
}
\value{
Text string with phrase
}
\description{
Very useful when you are feeling sad. Use it to brighten your day up
}
\examples{
praise(name = "Fairy")
}
\author{
Yash
}
