% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseME.R
\name{praiseME}
\alias{praiseME}
\title{A function that praises the user}
\usage{
praiseME(x)
}
\arguments{
\item{ali}{}
}
\value{
A string with the praise.
}
\description{
\code{praiseME} returns a praise incorporating a user-speciefied string.
}
\examples{

praiseME("Ali")
}
