% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseme.r
\name{praiseme}
\alias{praiseme}
\title{Print a short message of praise to the user}
\usage{
praiseme(praisefor)
}
\arguments{
\item{praisefor}{The verb or subject to praise the user for, entered as a string. If blank, a generic praise message will be returned.}
}
\value{
A short praise message as a string.
}
\description{
Takes a string as input and returns a short message of praise using the string.
}
\examples{
praiseme()
praiseme("art")
praiseme("coding")
}
