% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseme.R
\name{praiseme}
\alias{praiseme}
\title{A function that praises the user}
\usage{
praiseme(your_name)
}
\arguments{
\item{your_name}{A string that is your name.}
}
\value{
A string with the praise
}
\description{
\code{praiseme} returns praise incorporating a user-specified string.
}
\examples{

praiseme("John")
}
